Substantial cleanup + documentation pass
