+++ purescript.cabal
@@ -240,6 +240,7 @@ library
     Language.PureScript.CoreFn.Binders
     Language.PureScript.CoreFn.CSE
     Language.PureScript.CoreFn.Desugar
+    Language.PureScript.CoreFn.Desugar.Utils
     Language.PureScript.CoreFn.Expr
     Language.PureScript.CoreFn.FromJSON
     Language.PureScript.CoreFn.Meta
